Moved CustomUser to users app. Refactor code for Custom User and users must log in with email only.
